Add files via upload
The CPU will automatically use both the min and max inputted by the player, the player's guesses, and their own guesses to learn and sort through the possible range of values the number could be. The optimal strategy in this game is to typically halve the possible set of numbers with each guess. (ex if the range is 1-100, I guess 50, and if that's too high, I would then guess 25), so I implemented that strategy into the CPU. Now its a race between the CPU and player to who can uncover the random number first.